refactor: cleanup map generation
