+++ src/game/map/generation/location-types.ts
